renterhost: Rename package to rhp
